breachwatch_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
PyYAML
# For potential future Celery integration
# celery
# redis
